+++ package.json
@@ -1,44 +1,21 @@
 {
-  "name": "swtl",
-  "version": "0.3.0",
-  "description": "",
-  "main": "index.js",
+  "name": "@swtl/root",
+  "private": true,
+  "license": "MIT",
   "type": "module",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/thepassle/swtl.git"
+  },
   "scripts": {
-    "dev": "esbuild ./demo/sw.js --bundle --outfile=./demo/bundled-sw.js --watch --format=iife --servedir=demo",
-    "start": "node --watch dev.js",
-    "test": "node --watch --test tests/tests.js"
+    "test": "npm run test --workspaces"
   },
-  "exports": {
-    ".": "./index.js",
-    "./html.js": "./html.js",
-    "./await.js": "./await.js",
-    "./render.js": "./render.js",
-    "./router.js": "./router.js",
-    "./strategies.js": "./strategies.js",
-    "./slot.js": "./slot.js",
-    "./ssr/*.js": "./ssr/*.js"
-  },
-  "files": [
-    "html.js",
-    "render.js",
-    "router.js",
-    "symbol.js",
-    "await.js",
-    "strategies.js",
-    "index.js",
-    "slot.js",
-    "ssr/*"
-  ],
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "dependencies": {},
   "devDependencies": {
     "esbuild": "^0.19.0",
     "lit": "^3.1.2"
   },
-  "dependencies": {
-    "@lit-labs/ssr": "^3.2.2",
-    "@lit-labs/ssr-dom-shim": "^1.2.0"
-  }
+  "workspaces": [
+    "packages/*"
+  ]
 }
