+++ packages/core/demo/bundled-sw.js
