+++ packages/core/demo/pages/HtmlPage.js
