+++ packages/core/render.js
