+++ packages/core/slot.js
