+++ packages/core/strategies.js
