+++ packages/core/symbol.js
