+++ packages/core/test/html.test.js
@@ -1,14 +1,7 @@
 import assert from 'node:assert';
 import { describe, it } from 'node:test';
-import { LitElement, css, html as litHtml } from 'lit';
 import { html } from '../html.js';
-import { renderToString } from '../render.js';
-import { litRenderer } from '../ssr/lit.js';
-import { Await, when } from '../await.js';
-import { Slot } from '../slot.js';
 import { COMPONENT_SYMBOL, CUSTOM_ELEMENT_SYMBOL } from '../symbol.js';
-import { Router } from '../router.js';
-import { defaultRenderer } from '../ssr/default.js';
 
 function Foo() {}
 function Bar() {
@@ -438,304 +431,5 @@ describe('children', () => {
           kind: COMPONENT_SYMBOL
         },
       ]);
-  });
-});
-
-describe('renderToString', () => {
-  function Foo() {
-    return html`<h1>foo</h1>`
-  }
-
-  function Bar({children}) {
-    return html`<h1>${children}</h1>`
-  }
-
-  function Baz({children}) {
-    return html`<h2>${children}</h2>`
-  }
-
-  it('basic', async () => {
-    const result = await renderToString(html`<h1>hello</h1>`);
-    assert.equal(result, '<h1>hello</h1>');
-  });
-
-  it('expressions', async () => {
-    const result = await renderToString(html`<h1>hello ${1}</h1>`);
-    assert.equal(result, '<h1>hello 1</h1>');
-  });
-
-  it('components', async () => {
-    const result = await renderToString(html`<${Foo}/>`);
-    assert.equal(result, '<h1>foo</h1>');
-  });
-
-  it('components children', async () => {
-    const result = await renderToString(html`<${Bar}>bar<//>`);
-    assert.equal(result, '<h1>bar</h1>');
-  });
-
-  it('falsey values', async () => {
-    const result = await renderToString(html`<h1>${0}${false}</h1>`);
-    assert.equal(result, '<h1>0false</h1>');
-  });
-
-  it('objects', async () => {
-    const result = await renderToString(html`<h1>${{a: 2}}</h1>`);
-    assert.equal(result, '<h1>{"a":2}</h1>');
-  });
-  
-  it('components nested children', async () => {
-    const result = await renderToString(html`<${Bar}><${Baz}>baz<//><//>`);
-    assert.equal(result, '<h1><h2>baz</h2></h1>');
-  });
-
-  it('generator', async () => {
-    function* generator() {
-      yield* html`<li>1</li>`;
-      yield* html`<li>2</li>`;
-    }
-
-    const result = await renderToString(html`<ul>${generator()}</ul>`);
-    assert.equal(result, '<ul><li>1</li><li>2</li></ul>');
-  });
-
-  it('stream', async () => {
-    const stream = new ReadableStream({
-      start(controller) {
-        ['a', 'b', 'c'].forEach(s => controller.enqueue(s));
-        controller.close();
-      }
-    });
-    const result = await renderToString(html`<ul>${stream}</ul>`);
-    assert.equal(result, '<ul>abc</ul>');
-  });
-
-  it('response', async () => {
-    const response = new Response('<h1>hello</h1>');
-    const result = await renderToString(html`<main>${response}</main>`);
-    assert.equal(result, '<main><h1>hello</h1></main>');
-  });
-
-  it('Component returning response', async () => {
-    function Foo() {
-      return new Response('hi');
-    }
-    const result = await renderToString(html`<main><${Foo}/></main>`);
-    assert.equal(result, '<main>hi</main>');
-  });
-
-  describe('custom element renderer', () => {
-    describe('default', () => {
-      it('basic', async () => {
-        const result = await renderToString(html`<my-el foo=1 bar>children</my-el>`);
-        assert.equal(result, '<my-el foo="1" bar>children</my-el>');
-      });
-
-      it('nested swtl component', async () => {
-        function Foo() {
-          return html`<h1>foo</h1>`
-        }
-        const result = await renderToString(html`<my-el foo=1 bar><${Foo}/></my-el>`);
-        assert.equal(result, '<my-el foo="1" bar><h1>foo</h1></my-el>');
-      });
-
-      it('nested custom element', async () => {
-        function Foo({children}) {
-          return html`<h1>${children}</h1>`
-        }
-        const result = await renderToString(html`<${Foo}><my-el></my-el><//>`);
-        assert.equal(result, '<h1><my-el></my-el></h1>');
-      });
-    });
-
-    describe('lit', () => {
-      class FooEl extends LitElement {
-        static styles = css`h1 { color: red; }`;
-        static properties = { 
-          disabled: { type: Boolean },
-          foo: { type: String },
-        };
-        render() {
-          return litHtml`<h1>foo: ${this.foo} ${this.disabled ? 'yyy' : 'zzz'}</h1>`;
-        }
-      }
-
-      customElements.define('foo-el', FooEl);
-
-      it('basic', async () => {
-        const result = await renderToString(html`<foo-el foo="bar" disabled>children</foo-el>`, [litRenderer]);
-        assert(result.includes('<foo-el>'));
-        assert(result.includes('<template shadowroot="open" shadowrootmode="open">'));
-        assert(result.includes('<style>h1 { color: red; }</style>'));
-        assert(result.includes('yyy'));
-        assert(result.includes('bar'));
-        assert(result.includes('children'));
-        assert(result.includes('</template>'));
-        assert(result.includes('</foo-el>'));
-      });
-
-      it('nested swtl component', async () => {
-        function Foo() {
-          return html`<h2>Foo</h2>`
-        }
-        const result = await renderToString(html`<foo-el foo="bar" disabled><${Foo}/></foo-el>`, [litRenderer]);
-        assert(result.includes('<foo-el>'));
-        assert(result.includes('<template shadowroot="open" shadowrootmode="open">'));
-        assert(result.includes('<style>h1 { color: red; }</style>'));
-        assert(result.includes('yyy'));
-        assert(result.includes('bar'));
-        assert(result.includes('<h2>Foo</h2>'));
-        assert(result.includes('</template>'));
-        assert(result.includes('</foo-el>'));
-      });
-
-      it('property', async () => {
-        class PropertyEl extends LitElement {
-          static properties = { foo: { type: Object } };
-          render() {
-            return litHtml`<h1>${this.foo.a}</h1>`;
-          }
-        }
-        customElements.define('property-el', PropertyEl);
-        const result = await renderToString(html`<property-el .foo=${{a: 'zzz'}}></property-el>`, [litRenderer]);
-        assert(result.includes('zzz'));
-      });
-    });
-
-    it('supports multiple renderers', async () => {
-      class A extends LitElement {
-        render() {
-          return litHtml`<h1>lit</h1>`
-        }
-      }
-      customElements.define('lit-el', A);
-
-      const result = await renderToString(html`<lit-el></lit-el><html-el></html-el>`, [litRenderer, defaultRenderer]);
-      assert(result.includes('<lit-el><template shadowroot'));
-      assert(result.includes('<html-el></html-el>'));
-    });
-  });
-
-//   // it('Async', async () => {
-//   //   const result = await renderToString(html`<${Async} task=${() => new Promise(r => setTimeout(() => r({foo: 'bar'}), 100))}>
-//   //     ${({state, data}) => html`
-//   //       ${when(state === 'pending', () => html`[PENDING]`)}
-//   //       ${when(state === 'success', () => html`[RESOLVED] ${data.foo}`)}
-//   //     `}
-//   //   <//>`);
-//   //   console.log(result);
-//   //   assert.equal(result, `<pending-task style="display: contents;" data-id="0">
-//   //   [PENDING]
-    
-//   // </pending-task>
-//   // <template data-id="0">
-    
-//   //   [RESOLVED] bar
-//   // </template>
-//   // <script>
-//   //   {
-//   //     let toReplace = document.querySelector('pending-task[data-id="0"]');
-//   //     const template = document.querySelector('template[data-id="0"]').content.cloneNode(true);
-//   //     toReplace.replaceWith(template);
-//   //   }
-//   // </script>`);
-//   // });
-
-  it('kitchensink', async () => {
-    function Html({children}) {
-      return html`<html><body>${children}</body></html>`;
-    }
-
-    function Foo({bar, baz}) {
-      return html`<h2>foo ${bar} ${baz}</h2>`
-    }
-
-    const result = await renderToString(html`<${Html}><h1>welcome ${1}</h1><${Foo} bar=${1} baz="2"/><footer>copyright</footer><//>`)
-    assert.equal(result, '<html><body><h1>welcome 1</h1><h2>foo 1 2</h2><footer>copyright</footer></body></html>');
-  }); 
-});
-
-describe('slots', () => {
-  function Default({slots}) {
-    return html`<h1>${slots.default}</h1>`;
-  }
-
-  function OnlyNamed({slots}) {
-    return html`<h1>${slots?.foo}</h1>`;
-  }
-
-  function DefaultAndNamed({slots}) {
-    return html`<h1>${slots?.default}</h1><h2>${slots?.foo}</h2>`;
-  }
-
-  function MultipleNamed({slots}) {
-    return html`<h1>${slots?.foo}</h1><h2>${slots?.bar}</h2>`;
-  }
-
-  it('default', async () => {
-    const result = await renderToString(html`<${Default}><${Slot}>hi<//><//>`);
-    assert.equal(result, '<h1>hi</h1>');  
-  });
-
-  it('default and named', async () => {
-    const result = await renderToString(html`<${DefaultAndNamed}><${Slot}>hi<//><${Slot} name="foo">foo<//><//>`);
-    assert.equal(result, '<h1>hi</h1><h2>foo</h2>');  
-  });
-
-  it('multiple named', async () => {
-    const result = await renderToString(html`<${MultipleNamed}><${Slot} name="foo">foo<//><${Slot} name="bar">bar<//><//>`);
-    assert.equal(result, '<h1>foo</h1><h2>bar</h2>');  
-  });
-  
-  it('only named', async () => {
-    const result = await renderToString(html`<${OnlyNamed}><${Slot} name="foo">foo<//><//>`);
-    assert.equal(result, '<h1>foo</h1>');  
-  });
-});
-
-
-globalThis.URLPattern = class URLPattern {
-  exec() {
-    return true;
-  }
-}
-
-describe('Router', () => {
-  it('creates a default response', async () => {
-    const router = new Router({
-      routes: [
-      {
-        path: '/foo',
-        render: () => html`foo`
-      }
-    ]});
-
-    const response = await router.handleRequest({url: 'https://example.com/foo'});
-    const result = await response.text();
-    assert.equal(result, 'foo');
-    
-    assert.equal(response.headers.get('Content-Type'), 'text/html');
-    assert.equal(response.headers.get('Transfer-Encoding'), 'chunked');
-  });
-
-  it('can override defaults via `options`', async () => {
-    const router = new Router({
-      routes: [
-      {
-        path: '/foo',
-        render: () => html`foo`,
-        options: {
-          status: 300,
-          headers: {
-            'Content-Type': 'foo'
-          }
-        }
-      }
-    ]});
-
-    const response = await router.handleRequest({url: 'https://example.com/foo'});
-    
-    assert.equal(response.status, 300);
-    assert.equal(response.headers.get('Content-Type'), 'foo');
   });
 });
