+++ packages/lit/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@swtl/lit",
+  "version": "0.1.2",
+  "description": "",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "test": "node --test test/*.test.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@lit-labs/ssr": "^3.2.2",
+    "@lit-labs/ssr-dom-shim": "^1.2.0",
+    "swtl": "^0.3.0"
+  },
+  "devDependencies": {}
+}
